sources/pipeline/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.4
coloredlogs==14.0
contourpy==1.2.1
croniter==3.0.3
cycler==0.12.1
dagster==1.8.1
dagster-duckdb==0.24.1
dagster-graphql==1.8.1
dagster-pipes==1.8.1
dagster-webserver==1.8.1
dnspython==1.16.0
docstring_parser==0.16
duckdb==1.0.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.64.3
grpcio-health-checking==1.62.3
h11==0.14.0
holidays==0.55
httptools==0.6.1
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.3
Jinja2==3.1.4
kaleido==0.2.1
kiwisolver==1.4.5
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.0.5
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
prophet==1.1.5
protobuf==4.25.4
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==3.11.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==73.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stanio==0.5.1
starlette==0.38.2
structlog==24.4.0
tabulate==0.9.0
tenacity==9.0.0
tomli==2.0.1
toposort==1.10
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.23.0
websockets==12.0
yarl==1.9.4
